+++ src/main.ts
@@ -7,6 +7,8 @@ import "./components/notebook";
 
 import "iframe-resizer/js/iframeResizer.contentWindow.js";
 import * as lithtml from "lit-html";
+import katex from "katex";
+import * as YAML from "yaml";
 
 declare global {
   interface Window {
@@ -14,6 +16,8 @@ declare global {
     html?: typeof lithtml.html;
     svg?: typeof lithtml.svg;
     litHtml?: typeof lithtml;
+    katex?: typeof katex;
+    YAML?: typeof YAML;
   }
 }
 
@@ -21,9 +25,11 @@ declare global {
 window.html = lithtml.html;
 window.svg = lithtml.svg;
 window.litHtml = lithtml;
+window.katex = katex;
+window.YAML = YAML;
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-// window.initialNotebookContent = require("./debugNotebooks/introNotebook.nb").default;
+// window.initialNotebookContent = require("./debugNotebooks/promises.nb").default;
 
 document.body.innerHTML += `
 <base target="_parent" />