+++ src/runtime/create.ts
@@ -27,6 +27,9 @@ import * as LitElement from "lit-element";
 import * as LitHtml from "lit-html";
 import MarkdownIt from "markdown-it";
 import { precompileJavascriptCode } from "../cellTypes/javascript/precompile";
+import katex from "katex";
+import * as YAML from "yaml";
+import { hookMarkdownItToKaTeX } from "../components/helpers/katex";
 
 declare const STARBOARD_NOTEBOOK_VERSION: string;
 
@@ -41,7 +44,7 @@ function getInitialContent() {
     return textToNotebookContent((notebookContentElement as HTMLElement).innerText);
   }
 
-  return { frontMatter: "", cells: [] };
+  return { cells: [], metadata: {} };
 }
 
 export function createRuntime(notebook: StarboardNotebookElement): Runtime {
@@ -124,7 +127,7 @@ export function createRuntime(notebook: StarboardNotebookElement): Runtime {
           let cellElement: CellElement | null = rt.dom.cells[0] || null;
 
           while(cellElement) {
-            if (opts.onlyRunOnLoad && !cellElement.cell.properties.runOnLoad) {
+            if (opts.onlyRunOnLoad && !cellElement.cell.metadata.properties.run_on_load) {
               // Don't run this cell..
             } else {
               await cellElement.run();
@@ -193,7 +196,7 @@ function createExports(): RuntimeExports {
       JavascriptEvaluator: JavascriptEvaluator,
       createCellProxy: createCellProxy,
       hookMarkdownItToPrismHighlighter: hookMarkdownItToPrismHighlighter,
-      
+      hookMarkdownItToKaTeX: hookMarkdownItToKaTeX,
       cellToText: cellToText,
       notebookContentToText: notebookContentToText,
       precompileJavascriptCode: precompileJavascriptCode,
@@ -206,6 +209,8 @@ function createExports(): RuntimeExports {
       LitElement: LitElement,
       LitHtml: LitHtml,
       MarkdownIt: MarkdownIt,
+      KaTeX: katex,
+      YAML: YAML,
     }
   };