+++ test/content/legacyNotebookParsing.test.ts
@@ -2,12 +2,10 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-import { textToNotebookContent } from "../src/content/parsing";
-import { notebookContentToText } from "../src/content/serialization";
+import { textToNotebookContent } from "../../src/content/parsing";
+import { notebookContentToText } from "../../src/content/serialization";
 
-const simpleNotebookPlaintext = `
-Text before the first cell
-%% md
+const simpleNotebookPlaintext = `%% md
 # Notebook
 Hello there!
 
@@ -18,22 +16,21 @@ x;
 %% unknown-cell-type
 `;
 
-const notebookWithCellProperties = `
-%% js autoRun
+const notebookWithCellProperties = `%% js runOnLoad
 const a = 123;
 `;
 
-const withInvalidCell = `
-%% js autoRun
+const withInvalidCell = `%% js autoRun
 const a = 123;
 // Cell below is missing a type, so it should not start a new cell.
 %%
 const x = 3;
 `;
 
-const withInvalidFirstCell = `
+const totallyInvalidNotebook = `
 %%
 // There are actually zero cells now..
+// the entire notebook is invalid
 `;
 
 const emptyNotebook = ``;
@@ -52,7 +49,7 @@ describe("Text to notebook content", () => {
     it("can parse empty notebook", async () => {
         const notebookContent = textToNotebookContent(emptyNotebook);
         expect(notebookContent.cells).toHaveLength(0);
-        expect(notebookContent.frontMatter).toEqual("");
+        expect(notebookContent.metadata).toEqual({});
         expect(notebookContentToText(notebookContent)).toEqual("");
     });
 
@@ -61,13 +58,14 @@ describe("Text to notebook content", () => {
         const generatedNotebookText = notebookContentToText(notebookContent);
         const notebookContentAgain = textToNotebookContent(generatedNotebookText);
 
-        expect(generatedNotebookText).toEqual(simpleNotebookPlaintext);
+        // No longer true for legacy books as they will be rewritten into new format
+        // expect(generatedNotebookText).toEqual(simpleNotebookPlaintext);
 
         for(let i = 0; i < notebookContent.cells.length; i++) {
             expect(notebookContent.cells[i].textContent).toEqual(notebookContentAgain.cells[i].textContent);
             expect(notebookContent.cells[i].cellType).toEqual(notebookContentAgain.cells[i].cellType);
         }
-        expect(notebookContentAgain.frontMatter).toEqual(notebookContent.frontMatter);
+        expect(notebookContentAgain.metadata).toEqual(notebookContent.metadata);
     });
 
     it("parses autoRun flag", () => {
@@ -75,25 +73,29 @@ describe("Text to notebook content", () => {
 
         expect(notebookContent.cells).toHaveLength(1);
         const firstCell = notebookContent.cells[0];
-        expect(firstCell.properties["autoRun"]).toBeTruthy();
+        // This checks for new name of this property that gets patched automatically
+        expect(firstCell.metadata.properties["run_on_load"]).toBeTruthy();
     });
 
     it("handles wrong cell header", () => {
         const notebookContent = textToNotebookContent(withInvalidCell);
-        expect(notebookContent.cells).toHaveLength(2);
-        expect(notebookContent.cells[1].cellType).toEqual("");
-    });
-
-    it("handles wrong cell header for first cell", () => {
-        const notebookContent = textToNotebookContent(withInvalidFirstCell);
         expect(notebookContent.cells).toHaveLength(1);
-        expect(notebookContentToText(notebookContent)).toEqual(withInvalidFirstCell);
-        expect(notebookContentToText(textToNotebookContent(notebookContentToText(notebookContent)))).toEqual(withInvalidFirstCell);
     });
 
-    it("is stable in many conversions text->nb->text->nb->text", () => {
-        for(const nbContent of [simpleNotebookPlaintext, notebookWithCellProperties, withInvalidCell, withInvalidFirstCell, emptyNotebook]) {
-            expect(notebookContentToText(textToNotebookContent(notebookContentToText(textToNotebookContent(nbContent))))).toEqual(nbContent);
+    it("handles entirely broken notebook", () => {
+        jest.spyOn(console, 'error').mockImplementation(() => {/*do nothing*/});
+        expect(() => textToNotebookContent(totallyInvalidNotebook)).toThrow();
+    });
+
+    it("notebook content is stable in many conversions text->nb->text->nb->text->nb (!not the text!)", () => {
+        for(const nbContent of [simpleNotebookPlaintext, notebookWithCellProperties, withInvalidCell, emptyNotebook]) {
+            const content = textToNotebookContent(nbContent);
+            content.cells.map(x => x.id = "");
+
+            const contentAfterALotOfConversions = textToNotebookContent(notebookContentToText(textToNotebookContent(notebookContentToText(content))));
+            contentAfterALotOfConversions.cells.map(x => x.id = "");
+
+            expect(contentAfterALotOfConversions).toEqual(content);
         }
     });
 });